* adding fraud project with sln correctly
