+++ ExampleProjects/FraudDetection/NeuralNetworks.Examples.FraudDetection.Web/.bowerrc
