+++ ExampleProjects/FraudDetection/NeuralNetworks.Examples.FraudDetection.Web/Pages/About.cshtml.cs
@@ -2,16 +2,17 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 
-namespace NeuralNetworks.Examples.FraudDetection.Pages
+namespace NeuralNetworks.Examples.FraudDetection.Web.Pages
 {
-    public class IndexModel : PageModel
+    public class AboutModel : PageModel
     {
+        public string Message { get; set; }
+
         public void OnGet()
         {
-
+            Message = "Your application description page.";
         }
     }
 }
