+++ ExampleProjects/FraudDetection/NeuralNetworks.Examples.FraudDetection.Web/Pages/Contact.cshtml.cs
@@ -4,15 +4,15 @@ using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 
-namespace NeuralNetworks.Examples.FraudDetection.Pages
+namespace NeuralNetworks.Examples.FraudDetection.Web.Pages
 {
-    public class AboutModel : PageModel
+    public class ContactModel : PageModel
     {
         public string Message { get; set; }
 
         public void OnGet()
         {
-            Message = "Your application description page.";
+            Message = "Your contact page.";
         }
     }
 }
