+++ ExampleProjects/FraudDetection/NeuralNetworks.Examples.FraudDetection.Web/Pages/Index.cshtml
