+++ ExampleProjects/FraudDetection/NeuralNetworks.Examples.FraudDetection.Web/Pages/Index.cshtml.cs
@@ -2,17 +2,16 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 
-namespace NeuralNetworks.Examples.FraudDetection.Pages
+namespace NeuralNetworks.Examples.FraudDetection.Web.Pages
 {
-    public class ContactModel : PageModel
+    public class IndexModel : PageModel
     {
-        public string Message { get; set; }
-
         public void OnGet()
         {
-            Message = "Your contact page.";
+
         }
     }
 }
