+++ ExampleProjects/FraudDetection/NeuralNetworks.Examples.FraudDetection.Web/Pages/_Layout.cshtml
@@ -3,7 +3,7 @@
 <head>
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>@ViewData["Title"] - NeuralNetworks.Examples.FraudDetection</title>
+    <title>@ViewData["Title"] - NeuralNetworks.Examples.FraudDetection.Web</title>
 
     <environment include="Development">
         <link rel="stylesheet" href="~/lib/bootstrap/dist/css/bootstrap.css" />
@@ -26,7 +26,7 @@
                     <span class="icon-bar"></span>
                     <span class="icon-bar"></span>
                 </button>
-                <a asp-page="/Index" class="navbar-brand">NeuralNetworks.Examples.FraudDetection</a>
+                <a asp-page="/Index" class="navbar-brand">NeuralNetworks.Examples.FraudDetection.Web</a>
             </div>
             <div class="navbar-collapse collapse">
                 <ul class="nav navbar-nav">
@@ -41,7 +41,7 @@
         @RenderBody()
         <hr />
         <footer>
-            <p>&copy; 2017 - NeuralNetworks.Examples.FraudDetection</p>
+            <p>&copy; 2017 - NeuralNetworks.Examples.FraudDetection.Web</p>
         </footer>
     </div>
