+++ ExampleProjects/FraudDetection/NeuralNetworks.Examples.FraudDetection.Web/Pages/_ValidationScriptsPartial.cshtml
