+++ ExampleProjects/FraudDetection/NeuralNetworks.Examples.FraudDetection.Web/Pages/_ViewImports.cshtml
@@ -1,0 +1,3 @@
+@using NeuralNetworks.Examples.FraudDetection.Web
+@namespace NeuralNetworks.Examples.FraudDetection.Web.Pages
+@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
