+++ ExampleProjects/FraudDetection/NeuralNetworks.Examples.FraudDetection.Web/Pages/_ViewStart.cshtml
