+++ ExampleProjects/FraudDetection/NeuralNetworks.Examples.FraudDetection.Web/appsettings.Development.json
