+++ ExampleProjects/FraudDetection/NeuralNetworks.Examples.FraudDetection.Web/appsettings.json
