+++ ExampleProjects/FraudDetection/NeuralNetworks.Examples.FraudDetection.Web/bower.json
