+++ ExampleProjects/FraudDetection/NeuralNetworks.Examples.FraudDetection.Web/bundleconfig.json
