+++ ExampleProjects/FraudDetection/NeuralNetworks.Examples.FraudDetection.Web/wwwroot/js/site.js
