+++ ExampleProjects/FraudDetection/NeuralNetworks.Examples.FraudDetection.Web/wwwroot/lib/bootstrap/.bower.json
