+++ ExampleProjects/FraudDetection/NeuralNetworks.Examples.FraudDetection.Web/wwwroot/lib/jquery-validation/.bower.json
