+++ ExampleProjects/FraudDetection/NeuralNetworks.Examples.FraudDetection.Web/wwwroot/lib/jquery/.bower.json
