+++ ExampleProjects/FraudDetection/NeuralNetworks.Examples.FraudDetection.Web/wwwroot/lib/jquery/LICENSE.txt
