+++ ExampleProjects/FraudDetection/NeuralNetworks.Examples.FraudDetection.sln
@@ -1,0 +1,34 @@
+
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.26124.0
+MinimumVisualStudioVersion = 15.0.26124.0
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NeuralNetworks.Examples.FraudDetection.Web", "NeuralNetworks.Examples.FraudDetection.Web\NeuralNetworks.Examples.FraudDetection.Web.csproj", "{63781E3C-665E-45C4-8409-D716F82D67F0}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|Any CPU = Release|Any CPU
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{63781E3C-665E-45C4-8409-D716F82D67F0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{63781E3C-665E-45C4-8409-D716F82D67F0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{63781E3C-665E-45C4-8409-D716F82D67F0}.Debug|x64.ActiveCfg = Debug|x64
+		{63781E3C-665E-45C4-8409-D716F82D67F0}.Debug|x64.Build.0 = Debug|x64
+		{63781E3C-665E-45C4-8409-D716F82D67F0}.Debug|x86.ActiveCfg = Debug|x86
+		{63781E3C-665E-45C4-8409-D716F82D67F0}.Debug|x86.Build.0 = Debug|x86
+		{63781E3C-665E-45C4-8409-D716F82D67F0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{63781E3C-665E-45C4-8409-D716F82D67F0}.Release|Any CPU.Build.0 = Release|Any CPU
+		{63781E3C-665E-45C4-8409-D716F82D67F0}.Release|x64.ActiveCfg = Release|x64
+		{63781E3C-665E-45C4-8409-D716F82D67F0}.Release|x64.Build.0 = Release|x64
+		{63781E3C-665E-45C4-8409-D716F82D67F0}.Release|x86.ActiveCfg = Release|x86
+		{63781E3C-665E-45C4-8409-D716F82D67F0}.Release|x86.Build.0 = Release|x86
+	EndGlobalSection
+EndGlobal
