--- ExampleProjects/NeuralNetworks.Examples.FraudDetection/Pages/_ViewImports.cshtml
@@ -1,3 +1,0 @@
-@using NeuralNetworks.Examples.FraudDetection
-@namespace NeuralNetworks.Examples.FraudDetection.Pages
-@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
